간단한 train, test 동작 구현 및 확인

+++ dluid-ai/src/main/java/org/kok202/dluid/ai/network/NeuralNetLayer.java
@@ -9,6 +9,6 @@ import org.deeplearning4j.nn.api.Layer;
 @Builder
 @AllArgsConstructor
 public class NeuralNetLayer {
-    long sequence;
+    int sequence;
     Layer layer;
 }
